# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de la API",
        "description": "Documentación de la API de la aplicación",
        "version": "1.0"
    },
    "paths": {
...
        "/api/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Obtener todas las reservas",
                "operationId": "f10e68ff0ce263ef754b2f640eda8d93",
                "responses": {
                    "200": {
                        "description": "Lista de reservas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Reservation"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Crear una reserva",
                "operationId": "4f59911f11563f79b4f01bd02222607d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateReservationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Error de conflicto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reservations/{id}": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Obtener una reserva",
                "operationId": "14c077250033bb10a7305f9724f10bd4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reserva no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Cancelar una reserva",
                "operationId": "3f3a1d6e7ffee88b7ef58128a7f9295c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva cancelada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reserva no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Actualizar una reserva",
                "operationId": "7b80201e204732d789095d85fe3bc71d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateReservationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reserva actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Error de conflicto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reserva no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateReservationRequest": {
                "title": "Create Reservation Request",
                "description": "Esquema de validación para crear una reserva",
                "required": [
                    "resource_id",
                    "reserved_at",
                    "duration"
                ],
                "properties": {
                    "resource_id": {
                        "description": "ID del recurso a reservar",
                        "type": "integer",
                        "example": 1
                    },
                    "reserved_at": {
                        "description": "Fecha y hora de la reserva en el formato d-m-Y H:i",
                        "type": "string",
                        "format": "date-time",
                        "example": "25-12-2023 14:30"
                    },
                    "duration": {
                        "description": "Duración de la reserva en minutos, debe ser al menos 1",
                        "type": "integer",
                        "example": 20
                    }
                },
                "type": "object"
            },
...
            "UpdateReservationRequest": {
                "title": "UpdateReservationRequest",
                "description": "Esquema de validación para actualizar una reserva",
                "required": [
                    "status",
                    "resource_id",
                    "reserved_at",
                    "duration"
                ],
                "properties": {
                    "status": {
                        "description": "Estado de la reserva. Solo se permite el valor 'confirmed'.",
                        "type": "string",
                        "enum": [
                            "confirmed"
                        ]
                    },
                    "resource_id": {
                        "description": "ID del recurso asociado a la reserva. Debe existir en la tabla 'resources'.",
                        "type": "integer"
                    },
                    "reserved_at": {
                        "description": "Fecha y hora de la reserva en el formato 'd-m-Y H:i'.",
                        "type": "string",
                        "format": "date-time"
                    },
                    "duration": {
                        "description": "Duración de la reserva en minutos. Debe ser al menos 1.",
                        "type": "integer",
                        "minimum": 1
                    }
                },
                "type": "object"
            },
...
            "Reservation": {
                "title": "Reservación",
                "description": "Modelo de reservación",
                "required": [
                    "resource_id",
                    "user_id",
                    "reserved_at",
                    "duration",
                    "status"
                ],
                "properties": {
                    "resource_id": {
                        "description": "ID del recurso reservado",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "ID del usuario que hizo la reserva",
                        "type": "integer"
                    },
                    "reserved_at": {
                        "description": "Fecha y hora de la reserva",
                        "type": "string",
                        "format": "date-time"
                    },
                    "duration": {
                        "description": "Duración de la reserva en minutos",
                        "type": "integer"
                    },
                    "status": {
                        "description": "Estado de la reserva",
                        "type": "string",
                        "enum": [
                            "confirmed",
                            "pending",
                            "cancelled"
                        ]
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}